# Conflicts:
#	crates/nana-ui-runtime/src/framework/tests.rs
#	crates/nana-ui-runtime/src/framework/text_input.rs
#	crates/nana-ui-runtime/src/world.rs
#	crates/nana-ui-runtime/src/world/accessibility.rs
#	crates/nana-ui-runtime/src/world/accessibility_viewport_tests.rs
#	crates/nana-ui-runtime/src/world/hit_test.rs
#	crates/nana-ui-runtime/src/world/mutation.rs
#	crates/nana-ui-scene/src/runtime_document.rs
#	crates/nana-ui-scene/src/scene/order.rs
#	crates/nana-ui-scene/src/scene/tests.rs

+++ crates/nana-ui-scene/src/scene/primitives/controls.rs
@@ -53,7 +53,7 @@ pub(super) fn build(context: &GeometryPaintContext<'_>, emit: &mut impl FnMut(Sc
             for (index, (_, label, _)) in rows.iter().enumerate() {
                 emit(component_text_primitive(
                     id,
-                    40u8.saturating_add(index as u8),
+                    40u64.saturating_add(index as u64),
                     label,
                     TextHorizontalAlignment::Start,
                     true,
